Renamed CollateralDetailsTable.tsx
