+++ src/components/vaults/[vaultid]/VaultCollateralDetailsTable.tsx
