+++ src/pages/vaults/[vaultid].tsx
@@ -8,7 +8,7 @@ import {
 import { Container } from '@components/commons/Container'
 import { VaultHeading } from '@components/vaults/[vaultid]/VaultHeading'
 import { VaultDetailsTable } from '@components/vaults/[vaultid]/VaultDetailsTable'
-import { CollateralDetails } from '@components/vaults/[vaultid]/CollateralDetailsTable'
+import { CollateralDetails } from '@components/vaults/[vaultid]/VaultCollateralDetailsTable'
 import { VaultLoansTable } from '@components/vaults/[vaultid]/VaultLoansTable'
 import { getWhaleApiClient } from '@contexts/WhaleContext'
 import { isAlphanumeric } from '../../utils/commons/StringValidator'
